Fix call process twice for "operation would block"
The processor.process(i_proto, o_proto) was being called in an infinite
loop and breaking on the first error in the server.listen() call.

The first response was an Ok(()) which is what is expected for a ping
call. Since it would not break, the next call would be empty and gets
and error
```
TransportError { kind: Unknown, message: "operation would block" }
```
This error is not very helpful but goes away by removing the loop { .. }
call.